feat(@angular-devkit/core): move AJV schema validation to core
