+++ packages/angular_devkit/core/package.json
@@ -11,6 +11,7 @@
     "core"
   ],
   "dependencies": {
+    "ajv": "~5.5.1",
     "chokidar": "^1.7.0",
     "source-map": "^0.5.6"
   }
