--- packages/angular_devkit/core/src/json/schema/serializers/templates/javascript/prop-any.ejs
@@ -1,14 +1,0 @@
-{
-  value: undefined,
-  get() { return this.value === undefined ? <%= 'default' in schema ? JSON.stringify(schema.default) : 'undefined' %> : this.value; },
-  set(v) {
-    if (v === undefined && <%= !required %>) {
-      this.value = undefined;
-      return;
-    }
-    this.value = v;
-  },
-  isDefined() { return this.value !== undefined; },
-  remove() { this.set(undefined); },
-  schema() { return <%= JSON.stringify(schema) %>; },
-}
