--- packages/angular_devkit/core/src/json/schema/serializers/templates/javascript/root.ejs
@@ -1,20 +1,0 @@
-<%
-const extras = {
-  exceptions,
-  options,
-  strings,
-  symbols,
-  templates,
-};
-%>
-
-const holder = <%= templates.subschema(Object.assign({
-  name: name,
-  path: '',
-  required: false,
-  schema: schema,
-}, extras))
-%>;
-
-holder.set(value);
-return holder.get();
