--- packages/angular_devkit/core/src/json/schema/serializers/templates/javascript/subschema.ejs
@@ -1,38 +1,0 @@
-<%
-const extras = {
-  exceptions,
-  options,
-  path: (path ? path + '/' : '') + name,
-  strings,
-  symbols,
-  templates,
-};
-
-if (!schema) {
-  %>null<%
-} else if (schema === true) {
-%><%=
-  templates.prop_any(Object.assign({
-    name,
-    required,
-    schema: {},
-  }, extras))
-%><%
-} else if (!('type' in schema)) {
-%><%=
-  templates.prop_any(Object.assign({
-    name,
-    required,
-    schema: {},
-  }, extras))
-%><%
-} else {
-%><%=
-  templates['prop_' + schema.type](Object.assign({
-    name: name,
-    required,
-    schema: schema,
-  }, extras))
-%><%
-}
-%>
