+++ packages/angular_devkit/schematics/package.json
@@ -17,8 +17,7 @@
   ],
   "dependencies": {
     "@angular-devkit/core": "0.0.0",
-    "@ngtools/json-schema": "^1.1.0",
-    "ajv": "~5.5.1"
+    "@ngtools/json-schema": "^1.1.0"
   },
   "peerDependencies": {
     "rxjs": "^5.5.2"
