+++ packages/angular_devkit/schematics/tools/index.ts
@@ -14,5 +14,4 @@ export { FileSystemEngineHost } from './file-system-engine-host';
 export { NodeModulesEngineHost } from './node-module-engine-host';
 export { NodeModulesTestEngineHost } from './node-modules-test-engine-host';
 
-export { AjvSchemaRegistry } from './ajv-option-transform';
 export { validateOptionsWithSchema } from './schema-option-transform';
